Split off colvarproxy_system class
